Mount datasets on RCNFS only.
Previously datasets were mounted on all nodes in the cluster. But this
doesn't work when you want to write something to the dataset, because
there is no central coordination for the writes to the remote blockstore
among the nodes. Instead, I'm going to mount the datasets on RCNFS only
(and later export them to all the other nodes in the cluster via NFS).