Moved NFS off control network to exp network.
NFS was previously on the control network, which ended up flooding the
control network under high load. Moved the NFS network to the experiment
network.

One concern I have about this, however, is how this will affect RAMCloud
and TorcDB experiments. This remains to be seen, but for now I need to
stay off of the control network.